Added columns to the components list to display the quantity.
To improve efficiency during the SMT assembly process, I have added a new column in the component list to display the quantity. Users can enable or disable the display of this column using checkboxes.

However, there are the following unresolved issues:

Unable to determine if the quantity of PCBs has been modified for automatic updating of the list (related to modifying components associated with reference markers). I lack knowledge in JavaScript and don't know how to retrieve HTML element IDs. It appears that passing parameters is necessary, but implementing this change would require significant modifications to existing code. I hope the author can suggest a better approach for retrieving the parameters.
After adding the extra column, the overall width of the original list may be insufficient. Although it's possible to slightly increase the list's width, I have refrained from making modifications to maintain code consistency.